+++ middleware/service/include/service/manager/state.h
@@ -264,21 +264,33 @@ namespace casual
                         CASUAL_SERIALIZE( service);
                      )
                   };
-                  
+
+   
+                  struct Unset 
+                  {
+                     CASUAL_LOG_SERIALIZE(
+                        CASUAL_SERIALIZE( "Unset");
+                     )
+                  };
+
+                  // Directive for deadline changes, either a new time_point or Unset to 
+                  // unset timer ( i.e. no pending deadlines left)
+                  using Directive = std::variant< common::chronology::time_point, Unset>;
+
                } // deadline
 
                struct Deadline
                {      
                   std::optional< common::chronology::time_point> add( deadline::Entry entry);
-                  std::optional< common::chronology::time_point> remove( const common::strong::correlation::id& correlation);
-                  std::optional< common::chronology::time_point> remove( const std::vector< common::strong::correlation::id>& correlations);
+                  std::optional< deadline::Directive> remove( const common::strong::correlation::id& correlation);
+                  std::optional< deadline::Directive> remove( instance::sequential::id::type instance);
 
                   deadline::Entry* find_entry( const common::strong::correlation::id& correlation);
 
                   struct Expired
                   {
                      std::vector< deadline::Entry> entries;
-                     std::optional< common::chronology::time_point> deadline;
+                     std::optional< deadline::Directive> deadline;
 
                      CASUAL_LOG_SERIALIZE(
                         CASUAL_SERIALIZE( entries);
@@ -287,6 +299,8 @@ namespace casual
                   };
 
                   Expired expired( common::chronology::time_point now = common::chronology::time_point::clock::now());
+
+                  const auto& entries() const noexcept { return m_entries;}
 
                   CASUAL_LOG_SERIALIZE( 
                      CASUAL_SERIALIZE_NAME( m_entries, "entries");
@@ -501,11 +515,29 @@ namespace casual
             void remove_service( instance::sequential::id::type instance_id, service::id::type service_id);
             inline void remove_service( instance::concurrent::id::type, service::id::type) { /*no op*/}
 
+            //! @returns the sequential id for the given `pid`, or 'nil-id' if not found
+            instance::sequential::id::type find_sequential( common::strong::process::id pid) const;
+
             CASUAL_LOG_SERIALIZE(
                CASUAL_SERIALIZE( sequential);
                CASUAL_SERIALIZE( concurrent);
             )
          };
+
+         namespace remove
+         {
+            struct Result
+            {
+               std::vector< state::instance::Reservation> reservations;
+               //! potentially new deadline to be set.
+               std::optional< service::pending::deadline::Directive> deadline;
+
+               CASUAL_LOG_SERIALIZE(
+                  CASUAL_SERIALIZE( reservations);
+                  CASUAL_SERIALIZE( deadline);
+               )
+            };
+         } // remove
 
          enum struct Runlevel : short
          {
@@ -604,9 +636,9 @@ namespace casual
          
          //! removes the instance (deduced from `pid`) and remove the instance from all services 
          //! @returns possible reservations of the removed instance (in practice 0..1)
-         [[nodiscard]] std::vector< state::instance::Reservation> remove( common::strong::process::id pid);
+         [[nodiscard]] state::remove::Result remove( common::strong::process::id pid);
          //! @returns possible reservations of the removed instance
-         std::vector< state::instance::Reservation> remove( common::strong::ipc::id ipc);
+         [[nodiscard]] state::remove::Result remove( common::strong::ipc::id ipc);
 
          //! Tries to reserve a sequential instance for the given `service`
          //! @return id of the instance, or 'nil-id' if no idle is found