+++ documentation/changelog.md
@@ -1,6 +1,28 @@
 # Changelog
 This is the changelog for `casual` and all changes are listed in this document.
 
+## [1.8.6] - 2025-11-21
+
+### Fixes
+- event: casual-event-service-log should flush every line ([#640](https://github.com/casualcore/casual/issues/640))
+- service: unittest for no-reply reservations and deadlines
+- service: possible UB in remove deadline - deadline improvements
+- tools: call cli now uses the same correlation for call as lookup
+- common: argument doesn't suggest deprecated options during auto-complete
+- file: make sure to expand paths in order to catch some bad input ([#638](https://github.com/casualcore/casual/issues/638))
+- documentation: update build status in readme
+
+## [1.8.5] - 2025-11-07
+
+### Fixes
+- service: SM doesn't remove pending deadlines when an instance dies ([#631](https://github.com/casualcore/casual/issues/631))
+
+## [1.8.4] - 2025-11-04
+
+### Fixes
+- domain: if server exit during boot the task might get stuck ([#624](https://github.com/casualcore/casual/issues/624))
+
+
 ## [1.8.3] - 2025-10-06
 
 ### Fixes